cpcbproject/requirements.txt
amqp==5.1.1
arabic-reshaper==3.0.0
asgiref==3.5.2
asn1crypto==1.5.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
confusable-homoglyphs==3.2.0
cryptography==40.0.2
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.6
Django==4.1.2
django-admin-interface==0.26.0
django-ckeditor==6.5.1
django-colorfield==0.9.0
django-crispy-forms==2.0
django-filter==21.1
django-import-export==3.2.0
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-mathfilters==1.0.0
django-mptt==0.14.0
django-pdb==0.6.2
django-registration==2.4.1
django-smart-selects==1.6.0
django-taggit==3.0.0
django-widget-tweaks==1.4.8
dnspython==2.3.0
et-xmlfile==1.1.0
filelock==3.12.0
gprof2dot==2021.2.21
idna==3.4
iniconfig==1.1.1
kombu==5.2.4
MarkupPy==1.14
numpy==1.24.2
odfpy==1.4.1
openpyxl==3.1.2
oscrypto==1.3.0
pandas==1.5.3
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.1
Pillow==9.3.0
platformdirs==3.2.0
pluggy==1.0.0
prompt-toolkit==3.0.38
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pymongo==4.3.3
pyOpenSSL==23.1.1
pyparsing==3.0.9
pypdf==3.8.1
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.0
python-dotenv==0.12.0
python-slugify==8.0.1
pytz==2020.1
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.12
requests==2.28.2
six==1.16.0
sqlparse==0.4.1
static3==0.7.0
tablib==3.5.0
text-unidecode==1.3
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2022.7
Unipath==1.0
uritools==4.0.1
urllib3==1.26.15
vine==5.0.0
virtualenv==20.22.0
wcwidth==0.2.6
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
